08.12 - final refactorings (step 5 - refactor the orders feature) - create a orders package. move all the order related code to the orders package
