+++ src/main/java/com/jparkkennaby/store/orders/OrderItem.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.entities;
+package com.jparkkennaby.store.orders;
 
 import java.math.BigDecimal;
