+++ src/main/java/com/jparkkennaby/store/orders/OrderNotFoundException.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.orders;
 
 public class OrderNotFoundException extends RuntimeException {
     public OrderNotFoundException() {
