+++ src/main/java/com/jparkkennaby/store/orders/OrderService.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.services;
+package com.jparkkennaby.store.orders;
 
 import org.springframework.security.access.AccessDeniedException;
 import java.util.List;
@@ -6,10 +6,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.jparkkennaby.store.auth.AuthService;
-import com.jparkkennaby.store.dtos.OrderDto;
-import com.jparkkennaby.store.exceptions.OrderNotFoundException;
-import com.jparkkennaby.store.mappers.OrdersMapper;
-import com.jparkkennaby.store.repositories.OrderRepository;
 
 import lombok.AllArgsConstructor;
