+++ src/main/java/com/jparkkennaby/store/orders/OrdersMapper.java
@@ -1,10 +1,6 @@
-package com.jparkkennaby.store.mappers;
+package com.jparkkennaby.store.orders;
 
 import org.mapstruct.Mapper;
-
-import com.jparkkennaby.store.dtos.OrderDto;
-
-import com.jparkkennaby.store.entities.Order;
 
 @Mapper(componentModel = "spring") // so spring is made aware and can make beans at runtime
 public interface OrdersMapper {
