+++ src/main/java/com/jparkkennaby/store/orders/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.entities;
+package com.jparkkennaby.store.orders;
 
 public enum PaymentStatus {
     PENDING,
