+++ src/main/java/com/jparkkennaby/store/orders/ProductDto.java
@@ -1,11 +1,11 @@
-package com.jparkkennaby.store.dtos;
+package com.jparkkennaby.store.orders;
 
 import java.math.BigDecimal;
 
 import lombok.Data;
 
 @Data
-public class OrderProductDto {
+public class ProductDto {
     private Long id;
     private String name;
     private BigDecimal price;
