+++ src/main/java/com/jparkkennaby/store/payments/CheckoutService.java
@@ -8,8 +8,8 @@ import com.jparkkennaby.store.carts.CartEmptyException;
 import com.jparkkennaby.store.carts.CartNotFoundException;
 import com.jparkkennaby.store.carts.CartRepository;
 import com.jparkkennaby.store.carts.CartService;
-import com.jparkkennaby.store.entities.Order;
-import com.jparkkennaby.store.repositories.OrderRepository;
+import com.jparkkennaby.store.orders.Order;
+import com.jparkkennaby.store.orders.OrderRepository;
 
 import lombok.RequiredArgsConstructor;
