+++ src/main/java/com/jparkkennaby/store/services/SeedService.java
@@ -13,11 +13,11 @@ import org.springframework.stereotype.Service;
 import com.github.javafaker.Faker;
 import com.jparkkennaby.store.carts.CartRepository;
 import com.jparkkennaby.store.exceptions.DatabaseSeedingException;
+import com.jparkkennaby.store.orders.OrderRepository;
 import com.jparkkennaby.store.products.Category;
 import com.jparkkennaby.store.products.CategoryRepository;
 import com.jparkkennaby.store.products.Product;
 import com.jparkkennaby.store.products.ProductRepository;
-import com.jparkkennaby.store.repositories.OrderRepository;
 import com.jparkkennaby.store.users.Role;
 import com.jparkkennaby.store.users.User;
 import com.jparkkennaby.store.users.UserRepository;
